Minify layout CSS. See NOTE
NOTE I'm not sure what was going on here. This minifaction was a result
of the build process. I.e. I ran yarn build and the file was minified.
Not sure why it was not minified earlier, or if some env difference is
causing it to be minified now but not on Markus' computer.