Update singular page names
The singular label names should be unique... which is a bit of a UX
blunder on Netlify's part. When viewing the "qualifications" page it
makes sense to have a "new page" button, but when selecting from an
arbitrary UI from the "New" dropdown seeing multiple options for "Page"
is not helpful.

I.e. from the dropdown creation menu it was not clear what page
corresponded to what collection.