Run job listing cron job two hours earlier and don't force
I was worried that a force push would overwrite CMS changes. If there
really is a conflict, better that the job simply fail and the existing
changes remain. The job will get run again 24 hours later.

A more robust solution might be to find away to use the auto commit
action to do a "git pull --rebase" beforehand. Since job files are all
auto generated (don't edit them manually!) there shouldn't be any
conflicts. The only reason autocommit actually fails is a discrepancy
with the remote, which happens only within the small time window that
the action has to run.