Update lang config for taxonomies
We are _not_ using the translations in the taxonomy files. The
translations are all based in the files under i18n/*. As it stands the
multi-lang admin features were causing an issue where the users
preferred language would determine which string was used for the
taxonomy added to a post.

What _this_ meant is that the translation strings weren't working,
because they expect english as the base lang.