Update the jobs script
It seems that the underlying site JSON has changed shape.

Furthermore, we no longer use the highly error-prone JSON.parse and
instead use the Node VM to handle evaluation of the raw script tag.

This too could prove to be error prone, if their redux state includes
side effects. Being a state for hydration this is hopefully unlikely,
although they do include functions in their app state which is what
broke the script the last time (Well, that and their changing data
shape).